requirements.txt
fastapi
uvicorn
crewai
crewai-tools
python-dotenv
pypdf
pydantic
langchain-openai
langchain-google-genai   
